feat: Refatoração dos repositórios para usar sql.DB diretamente
- Os repositórios foram atualizados para usar sql.DB em vez de SQLStr personalizado.
- Alteração realizada para alinhar o código ao padrão de Clean Architecture, tornando os repositórios responsáveis pelas queries e o banco apenas pela conexão.
