+++ database/sql.go
@@ -1,4 +1,4 @@
-package sql
+package database
 
 import (
 	"context"
