+++ internal/Fornecedores/handler.go
@@ -3,7 +3,6 @@ package fornecedor
 import (
 	"encoding/json"
 	"lapasta/internal/models"
-	"log"
 	"net/http"
 	"strconv"
 )
@@ -13,8 +12,6 @@ type FornecedorHandler interface {
 	EditarFornecedor(w http.ResponseWriter, r *http.Request)
 	ListarFornecedores(w http.ResponseWriter, r *http.Request)
 	BuscarFornecedorPorCNPJouNome(w http.ResponseWriter, r *http.Request)
-	CriarPedido(w http.ResponseWriter, r *http.Request)
-	ListarPedidosPorFornecedor(w http.ResponseWriter, r *http.Request)
 	BuscarPedidosFornecedorPorDescricaoOuId(w http.ResponseWriter, r *http.Request)
 }
 
@@ -124,61 +121,6 @@ func (h *fornecedorHandler) BuscarFornecedorPorCNPJouNome(w http.ResponseWriter,
 	response := models.ResponseDefaultModel{
 		IsSuccess: true,
 		Data:      fornecedor,
-	}
-
-	w.Header().Set("Content-Type", "application/json; charset=utf-8")
-	json.NewEncoder(w).Encode(response)
-}
-
-func (h *fornecedorHandler) CriarPedido(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json; charset=utf-8")
-	var pedido models.PedidoFornecedor
-	response := models.ResponseDefaultModel{IsSuccess: true}
-
-	if err := json.NewDecoder(r.Body).Decode(&pedido); err != nil {
-		log.Printf("Erro ao decodificar pedido: %v", err)
-		http.Error(w, `{"isSuccess":false,"errorMessage":"Erro ao decodificar o pedido"}`, http.StatusBadRequest)
-		return
-	}
-
-	log.Printf("Pedido recebido: %+v", pedido)
-
-	if err := h.service.CriarPedidoFornecedor(&pedido); err != nil {
-		log.Printf("Erro ao criar pedido: %v", err)
-		http.Error(w, `{"isSuccess":false,"errorMessage":"Erro ao criar pedido"}`, http.StatusInternalServerError)
-		return
-	}
-
-	response.Data = pedido
-	w.WriteHeader(http.StatusCreated)
-	json.NewEncoder(w).Encode(response)
-}
-
-func (h *fornecedorHandler) ListarPedidosPorFornecedor(w http.ResponseWriter, r *http.Request) {
-	valor := r.URL.Query().Get("fornecedorId")
-	response := models.ResponseDefaultModel{IsSuccess: true}
-
-	if valor == "" {
-		response.IsSuccess = false
-		response.ErrorMessage = "ID do fornecedor é obrigatório"
-		w.WriteHeader(http.StatusBadRequest)
-	} else {
-		id, err := strconv.Atoi(valor)
-		if err != nil {
-			response.IsSuccess = false
-			response.ErrorMessage = "ID inválido"
-			w.WriteHeader(http.StatusBadRequest)
-		} else {
-			pedidos, err := h.service.ListarPedidosPorFornecedor(id)
-			if err != nil {
-				response.IsSuccess = false
-				response.ErrorMessage = "Erro ao listar pedidos"
-				w.WriteHeader(http.StatusInternalServerError)
-			} else {
-				response.Data = pedidos
-				w.WriteHeader(http.StatusOK)
-			}
-		}
 	}
 
 	w.Header().Set("Content-Type", "application/json; charset=utf-8")
