+++ internal/Fornecedores/service.go
@@ -9,8 +9,6 @@ type FornecedorService interface {
 	EditarFornecedor(f *models.Fornecedor) error
 	ListarFornecedores(page int) ([]models.Fornecedor, error)
 	BuscarFornecedorPorCNPJouNome(valor string) ([]models.Fornecedor, error)
-	CriarPedidoFornecedor(pedido *models.PedidoFornecedor) error
-	ListarPedidosPorFornecedor(fornecedorId int) ([]models.PedidoFornecedor, error)
 	BuscarPedidosFornecedorPorDescricaoOuId(fornecedorId int, valor string) ([]models.PedidoFornecedor, error) 
 
 }
@@ -38,13 +36,6 @@ func (s *fornecedorService) ListarFornecedores(page int) ([]models.Fornecedor, e
 
 func (s *fornecedorService) BuscarFornecedorPorCNPJouNome(valor string) ([]models.Fornecedor, error) {
 	return s.repo.BuscarFornecedorPorCNPJouNome(valor)
-}
-func (s *fornecedorService) CriarPedidoFornecedor(pedido *models.PedidoFornecedor) error {
-	return s.repo.CriarPedidoFornecedor(pedido)
-}
-
-func (s *fornecedorService) ListarPedidosPorFornecedor(fornecedorId int) ([]models.PedidoFornecedor, error) {
-	return s.repo.ListarPedidosPorFornecedor(fornecedorId)
 }
 func (s *fornecedorService) BuscarPedidosFornecedorPorDescricaoOuId(fornecedorId int, valor string) ([]models.PedidoFornecedor, error) {
 	return s.repo.BuscarPedidosFornecedorPorDescricaoOuId(fornecedorId, valor)
