+++ internal/Ponto/handler.go
@@ -1,4 +1,3 @@
-// handler/ponto.go
 package ponto
 
 import (
