+++ internal/models/funcionario_model.go
